(PUP-3644) Copy v1 to v3
This commit simply makes a copy of the v1 http handler code
to a new file called v3.rb; no code changes yet other than
changing the class name from `V1` to `V3`.  Doing this as
an isolated commit so that subsequent changes to the v3.rb
code will have a more meaningful diff in the VCS history.

It also copies the spec tests for v1 over to v3.